refactor: Actualizar tipo de datos en modelos

+++ src/models/permission.model.js
@@ -6,7 +6,7 @@ const Permission = sequelize.define(
   'Permission',
   {
     id: {
-      type: DataTypes.INTEGER(11),
+      type: DataTypes.INTEGER,
       primaryKey: true,
       autoIncrement: true,
     },
