+++ src/models/task.js
@@ -5,7 +5,7 @@ const Task = sequelize.define(
   'Task',
   {
     id: {
-      type: DataTypes.INTEGER(11),
+      type: DataTypes.INTEGER,
       primaryKey: true,
       autoIncrement: true,
     },
